Move ExpandableNotification
